+++ lib/sql_server_dts_ns.rb
@@ -1,5 +1,7 @@
+# coding: utf-8
+
 # SqlServerDts module
 module SqlServerDts
   # current version
-  VERSION = '0.0.1'
+  VERSION = '0.1.2'
 end